requirements.txt
fastapi
pydantic
sentence-transformers
newspaper3k
openai
python-dotenv
requests
uvicorn
playwright
